Extract methods to helper class
