+++ OsmIntegrator/Interfaces/IOsmRefresherHelper.cs
@@ -1,0 +1,21 @@
+using System.Threading;
+using System.Threading.Tasks;
+using System.Net.Http;
+using System.Collections.Generic;
+using OsmIntegrator.Database.Models;
+using OsmIntegrator.Database;
+using OsmIntegrator.Tools;
+
+namespace OsmIntegrator.Interfaces
+{
+  public interface IOsmRefresherHelper
+  {
+    Task<Osm> GetContent(HttpContent content);
+    Task<Osm> GetContent(HttpContent content, CancellationToken cancellationToken);
+
+    Task Refresh(DbTile tile, ApplicationDbContext dbContextm, Osm osmRoot);
+
+    Task Refresh(List<DbTile> tiles, ApplicationDbContext dbContext, Osm osmRoot);
+
+  }
+}
